refactor: modularize response complete notifications
