+++ src/entrypoints/background/responseCompleteNotification/audio.test.ts
@@ -7,7 +7,7 @@ import {
 import {
   playResponseCompleteNotificationAudio,
   resetResponseCompleteNotificationAudioForTest,
-} from './responseCompleteNotificationAudio'
+} from './audio'
 
 const {
   getContextsMock,
