+++ src/entrypoints/background/responseCompleteNotification/deepResearchState.ts
