+++ src/entrypoints/background/responseCompleteNotification/geminiResponseRequest.test.ts
