+++ src/entrypoints/background/responseCompleteNotification/geminiResponseRequest.ts
