+++ src/entrypoints/background/responseCompleteNotification/readiness.ts
@@ -1,0 +1,5 @@
+import type { NotificationReadiness } from '@/services/responseCompleteNotificationSettings'
+
+export function isSendableReadiness(readiness: NotificationReadiness): boolean {
+  return readiness === 'allowed' || readiness === 'allowed-but-system-unknown'
+}
